Move mocks into mock.go
